% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{analyze_locus}
\alias{analyze_locus}
\title{Extract Locus Information from VCF Data for HAD Format}
\usage{
analyze_locus(locus, remove_double_hets = FALSE)
}
\arguments{
\item{locus}{A character string of the locus information from a column of the VCF data frame.}

\item{remove_double_hets}{Logical for determining if double heterozygous loci should be treated as missing information.}
}
\value{
A numeric vector of the number of reads for each allele at heterozygous loci.
}
\description{
A function used to fill the HAD converted data frame with the number of reads on each allele for heterozygous loci. Not exported.
}
\examples{
\dontrun{analyze_locus("1/0:8:0,6,0,2", remove_double_hets=TRUE)}
}
