% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{analyze_locus_colony}
\alias{analyze_locus_colony}
\title{Extract Locus Information from VCF Data for COLONY Format}
\usage{
analyze_locus_colony(locus)
}
\arguments{
\item{locus}{A character string of the locus information from a column of the VCF data frame.}
}
\value{
A numeric vector of the number of reads for each allele at heterozygous loci.
}
\description{
A function used to fill the Colony converted data frame. Not exported.
}
\examples{
\dontrun{analyze_locus("1/0:8:0,6,0,2")}
}
