% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_metadata_lines}
\alias{count_metadata_lines}
\title{Count number of metadata lines in VCF files}
\usage{
count_metadata_lines(filename)
}
\arguments{
\item{filename}{A character string of the data file path.}
}
\value{
Returns a numeric of the number of metadata lines in VCF file.
}
\description{
Read in a Variant Calling Format file and count the number of metadata lines to skip over for read_VCF function. Not exported.
}
\details{
Throws an error if file cannot be found. Metadata lines are identified by checking if the lines starts with "##". This function requires reading in the entire file, so if you have a large file and know the number of metadata lines in that file, you can save some run time in the read_VCF, VCF2HAD, VCF2Ploidy, and VCF2Colony functions by entering the number of metadata lines in their "skip_lines" argument.
}
\examples{
\dontrun{count_metadata_lines("./inst/extdata/example.vcf")}
}
