% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{handle_vcf2ploidy_args}
\alias{handle_vcf2ploidy_args}
\title{Handle Arguments for vcf2ploidy from Shiny Input}
\usage{
handle_vcf2ploidy_args(input)
}
\value{
A list containing cleaned arguments for vcf2ploidy()
}
\description{
Convert text input values from vcf2ploidy_app() into appropriate vectors of values for vcf2ploidy arguments
}
