% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_example_data.R
\name{make_example_vcf_file}
\alias{make_example_vcf_file}
\title{Generate a basic example VCF file}
\usage{
make_example_vcf_file(
  save_path = "./generated_example.vcf",
  min_reads = 8,
  max_reads = 200,
  num_loci = 10,
  num_specimens = 5
)
}
\arguments{
\item{save_path}{the file path and file name to save the generated VCF file}

\item{min_reads}{the minimum of total number of reads per locus}

\item{max_reads}{the maximum of total number of reads per locus}

\item{num_loci}{the number of loci (rows) to generate in the file}

\item{num_specimens}{the number of specimen (columns) to generate in the file}
}
\description{
Generate a basic example VCF file
}
\details{
The generated file will hold no meaningful information, this function is here only for testing the converter function.
}
\examples{
\dontrun{make_example_vcf_file()}
\dontrun{make_example_vcf_file(save_path="./my_example.vcf")}
\dontrun{make_example_vcf_file(min_reads=24, max_reads=100)}
\dontrun{make_example_vcf_file(num_loci=100, num_specimens=4)}
}
