% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_example_data.R
\name{make_locus}
\alias{make_locus}
\title{Generate Random Locus Data}
\usage{
make_locus(min_reads = 4, max_reads = 200)
}
\arguments{
\item{min_reads}{the minimum number of allele reads at each locus}

\item{max_reads}{the maximum number of allele reads at each locus}
}
\value{
A character string of locus data
}
\description{
Generates heterozygous, homozygous, or missing locus data. Not exported.
}
\details{
Heterozygous or homozgous loci are given a random number of reads for each allele.
}
\examples{
\dontrun{make_locus()}
\dontrun{make_locus(min_reads=16, max_reads=100)}
}
