% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_example_data.R
\name{make_specimen_names}
\alias{make_specimen_names}
\title{Generate specimen names}
\usage{
make_specimen_names(n)
}
\arguments{
\item{n}{the number of specimen names to generate}
}
\value{
A character vector of specimen names
}
\description{
A very, very crude and confusing way to generate character strings of a number and a letter. Not exported.
}
\details{
The generated locus information will yield any meaningful information, this function is here only for testing the converter function.
}
\examples{
\dontrun{make_specimen_names(8)}
}
