% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_VCF}
\alias{read_VCF}
\title{Read in VCF files}
\usage{
read_VCF(filename, skip_lines = 10)
}
\arguments{
\item{filename}{A character string of the data file path.}

\item{skip_lines}{Number of metadata lines to skip before reading in the VCF file. Metadata lines typically start with "##". Default is 10.}
}
\value{
This function returns a data frame table of the VCF data with the metadata and the first 9 columns removed.
}
\description{
Read in a Variant Calling Format file.
}
\details{
Throws an error if file cannot be found.
}
\examples{
\dontrun{read_VCF("./inst/extdata/example.vcf")}

}
