% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{VCF2colony}
\alias{VCF2colony}
\title{Convert Raw VCF Data to Colony Format}
\usage{
VCF2colony(filename, skip_lines = 10, out_filename)
}
\arguments{
\item{filename}{A character string of the data file path.}

\item{skip_lines}{Number of metadata lines to skip before reading in the VCF file. Metadata lines typically start with "##". Default is 10.}

\item{out_filename}{A character string of the resulting converted file path.}
}
\description{
Convert tbl_df output from read_VCF() to a format that can be read in by Colony.
}
\examples{
\dontrun{VCF2colony(filename = "./inst/extdata/example.vcf", out_filename = "./example.txt")}

}
