% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{VCF2HAD}
\alias{VCF2HAD}
\title{Convert Raw VCF Data to HAD Format}
\usage{
VCF2HAD(filename, skip_lines = 10, remove_double_hets = FALSE)
}
\arguments{
\item{filename}{A character string of the data file path.}

\item{skip_lines}{Number of metadata lines to skip before reading in the VCF file. Metadata lines typically start with "##". Default is 10.}

\item{remove_double_hets}{Logical for determining if double heterozygous loci should be treated as missing information. Should fix issues with GBS2Ploidy falsely labeling triploids.}
}
\value{
A data frame in Heterozygous Allele Depth (HAD) format
}
\description{
Convert tbl_df output from read_VCF() to a Heterozygous Allele Depth (HAD) format, to be read in by GBS2Ploidy.
}
\examples{
\dontrun{VCF2HAD("./inst/extdata/example.vcf")}
\dontrun{VCF2HAD("./inst/extdata/example.vcf", remove_dobule_hets=TRUE)}

}
