% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{vcf2ploidy_app}
\alias{vcf2ploidy_app}
\title{vcf2ploidy Shiny Application}
\usage{
vcf2ploidy_app()
}
\description{
Open a shiny gadget for running vcf2ploidy functions
}
\details{
Drag and drop VCF file to run vcf2had(), vcf2ploidy(), or vcf2colony() on and select parameters in graphical interface
}
\examples{
\dontrun{vcf2ploidy_app()}

}
