% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vcf2colony}
\alias{vcf2colony}
\title{Convert Raw VCF Data to COLONY Format}
\usage{
vcf2colony(filename, skip_lines = NULL, out_filename)
}
\arguments{
\item{filename}{A character string of the data file path.}

\item{skip_lines}{A numeric of the number of metadata lines to skip over in the VCF file. If left null, metadata lines are skipped over automatically by the count_metadata_lines function. The count_metadata_lines function requires reading in the entire file, so if you have a large file and know the number of metadata lines in that file, you can save some run time by entering the number of metadata lines in this argument.}

\item{out_filename}{A character string of the resulting converted file path.}
}
\description{
Convert tbl_df output from read_vcf() to a format that can be read in by COLONY software.
}
\examples{
\dontrun{vcf2colony(filename = "./inst/extdata/example.vcf", out_filename = "./example.txt")}

}
