% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vcf2had}
\alias{vcf2had}
\title{Convert Raw VCF Data to HAD Format}
\usage{
vcf2had(filename, skip_lines = NULL, remove_double_hets = FALSE)
}
\arguments{
\item{filename}{A character string of the data file path.}

\item{skip_lines}{A numeric of the number of metadata lines to skip over in the VCF file. If left null, metadata lines are skipped over automatically by the count_metadata_lines function. The count_metadata_lines function requires reading in the entire file, so if you have a large file and know the number of metadata lines in that file, you can save some run time by entering the number of metadata lines in this argument.}

\item{remove_double_hets}{Logical for determining if double heterozygous loci should be treated as missing information. Should fix issues with gbs2ploidy falsely labeling triploids.}
}
\value{
A data frame in Heterozygous Allele Depth (HAD) format
}
\description{
Convert tbl_df output from read_vcf() to a Heterozygous Allele Depth (HAD) format, to be read in by gbs2ploidy.
}
\examples{
\dontrun{vcf2had("./inst/extdata/example.vcf")}
\dontrun{vcf2had("./inst/extdata/example.vcf", remove_dobule_hets=TRUE)}

}
