% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{VCF2Ploidy_APP}
\alias{VCF2Ploidy_APP}
\title{VCF2Ploidy Shiny Application}
\usage{
VCF2Ploidy_APP()
}
\description{
Open a shiny application for running the VCF2Ploidy function
}
\details{
Drag and drop VCF file to run VCF2Ploidy() on and select parameters in graphical interface
}
\examples{
\dontrun{VCF2Ploidy_APP()}

}
